added handling for precision, tested to work at least up to 10

+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jkuusist <jkuusist@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 11:19:32 by jkuusist          #+#    #+#             */
-/*   Updated: 2020/03/02 15:31:10 by jkuusist         ###   ########.fr       */
+/*   Updated: 2020/03/04 15:05:21 by jkuusist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ void				ft_lstdel(t_list **alst, void (*del)(void *, size_t));
 void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
-int					ft_pow(int base, unsigned int exp);
+long				ft_pow(int base, unsigned int exp);
 void				ft_strrev(char *str, int len);
 char				*ft_itoa_base(long num, int base);
 char				*ft_strupper(char *s);
